Refactor: loose coupling between goals scored and team object
Since a team object used in a competition (or more broadly in a 'season' or multiple seasons simulation) has a much bigger scope than a single match, it would be a mistake to continue having the goals scored in given matches as fields of team objects.

This refactoring makes match score part of the match object, like the rest of match stats.