Penalties dataset API: select by final outcome
There are three direct 'outcome types' identified in the current data set:

* Goalkeeper saves
* Hits at post
* Off target shots

For the first two outcome types, we need to define their final outcome; actually it is the only outcome we are currently interested in at this point for the simulation implementation. The off-target shots are both a direct and a final outcome, so no further querying is required.

For goal attempts, there is also a subsequent outcome which will be defined dynamically during runtime, within the now in-play, goal attempt scope.