Integrate penalty data in match engine execution flow
Initial commit. This is mostly refactoring regarding the match engine functionality itself, as corners are maintained as the only potential missed penalty outcome.

Every outcome case will be integrated incrementally, both for better and more convenient testing, and for facilitating the incremental development of changes that will be needed in the match engine itself.