Add option for not 'preprocessing' match data loaded in Dataset
The idea about preprocessing loaded data by default, and as the only option, was based on the goal of using the data directly in the match model. Currently, we are going for the more basic case of just loading match event data as a specific and 'constant' match representation, where obviously 'preprocessing' is an undesirable side effect.

By 'preprocessing', in the present case, I mean deducing the end time of a transition from the context of the match data; i.e. loading the start time of the next transition as the end time of the current one (where this is applicable and correct), as this is not always explicit in the transition data record. This is actually not false and very convenient, if we want to construct a match model from the data, but as mentioned above it can complicate the logic when using the data representation 'literally', as in the present case of displaying the match commentary.

In the future, the current approach may influence how the MPN engine will be rewritten, as the match model will be constructed from scratch based on the 'unprocessed' match data.