Load competition data dynamically from tournament id
This is the last domain change for now. The objective so far has been to have only the tournament id as reference, and all other data to be loaded based on that id (so that, for example, we do not hardcode the competition id as well). Of course, this is for running a tournament simulation; for repository retrieval 'demos', we still need to hardcode separate entity ids. The tournament id, will be hardcoded in the tournament client code, however its hardcoding simulates future implementations, where it will be specified, for example, from user input or 'gameplay' logic.