Remove UNIQUE constraint for team full name
This constraint is obviously still desired, however it will be enforced at the application level. The reason for removing this constraint is a hack of having multiple records per team to reflect different team 'historical' ratings. We thus keep things simple by using different keys for the same team across seasons, but using the same full name for reference to the same team. The uniqueness of the team is guaranteed at the application level by not having duplicate team participations (at any point we only run a particular competition for a particular season, defined as 'tournament' in the database and application, so we are not affected by the presence of duplicate references to the same team as long as the participants are set correctly).

In a more complex application, where, for example, maintaining a complete gameplay environment running across seasons would be required, we would go for a robust solution, such as separating 'historical' and 'current' data.